better handling no image found
