+++ LocalStack/ui/src/types/docker.ts
@@ -1,0 +1,16 @@
+export interface DockerImage {
+ Containers: number,
+ Created: number,
+ Id: string,
+ Labels: Labels,
+ ParentId: string,
+ RepoDigest: [string],
+ RepoTags: [string],
+ SharedSize: number,
+ Size: number,
+ VirtualSize: number,
+}
+
+interface Labels {
+  [key: string]: string,
+}
