+++ LocalStack/ui/src/types/global.ts
@@ -1,0 +1,1 @@
+export declare type Optional<T> = T | null | undefined;
